requirements.txt
jax>=0.2.5
jaxlib>=0.1.56
numpy>=1.18.5
tensorboardX>=2.1
dm-haiku>=0.0.2
python-box>=5.1.1
frozendict>=1.2
tqdm>=4.49.0
pytest>=6.1.1
seaborn>=0.11.0
matplotlib>=3.3.1
imageio>=2.9.0
oryx>=0.1.3
drawnow>=0.72.5

# If you want to run the PyTorch reference tests
# torch==1.7.0.dev20200912
# torchsearchsorted==1.1
